requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
fastapi>=0.95.2
uvicorn>=0.22.0